Revamp examples
Allows for easier to run and configure examples!
