+++ .gitignore
@@ -1,3 +1,5 @@
+examples/config.ini
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
