+++ examples/config.ini
@@ -1,0 +1,6 @@
+[DEFAULT]
+BYNDER_DOMAIN=
+BYNDER_CLIENT_ID=
+BYNDER_CLIENT_SECRET=
+BYNDER_REDIRECT_URI=
+BYNDER_ACCESS_TOKEN=
